Fix VRAM type casting in MemoryPool for proper initialization
This update ensures that the DedicatedVideoMemory value retrieved from the adapter description is explicitly cast to u64, enhancing type safety and preventing potential issues with memory handling. This change maintains the existing functionality while improving code clarity.

+++ nu_scaler_core/src/gpu/memory.rs
@@ -331,7 +331,7 @@ impl MemoryPool {
                         // Get adapter description with properly created descriptor
                         let mut desc = zeroed::<Dxgi::DXGI_ADAPTER_DESC1>();
                         if adapter.GetDesc1(&mut desc).is_ok() {
-                            dedicated_vram = desc.DedicatedVideoMemory;
+                            dedicated_vram = desc.DedicatedVideoMemory as u64;
                         }
 
                         // Get current usage with properly created memory info struct
